+++ src/utils/vector2i.cpp
@@ -4,6 +4,6 @@ Vector2::Vector2() : x(0), y(0){
 
 }
 
-Vector2::Vector2(int dx, int dy) : x(dx), y(dy){
+Vector2::Vector2(float dx, float dy) : x(dx), y(dy){
 
 }
